Added delete dialog in event_log, not functional
